requirements.txt
aiogram==3.4.1
aiohttp==3.9.1
feedparser==6.0.10
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
aiosqlite==3.4.1
